+++ config/ac_setter.go
@@ -1,9 +1,0 @@
-package config
-
-type AccessControlSetter struct {
-	ErrorHandler []*ErrorHandler `hcl:"error_handler,block"`
-}
-
-func (acs *AccessControlSetter) Set(ehConf *ErrorHandler) {
-	acs.ErrorHandler = append(acs.ErrorHandler, ehConf)
-}