+++ config/api.go
@@ -10,13 +10,14 @@ var _ Inline = &API{}
 
 // API represents the <API> object.
 type API struct {
+	ErrorHandlerSetter
 	AccessControl        []string  `hcl:"access_control,optional"`
 	BasePath             string    `hcl:"base_path,optional"`
 	CORS                 *CORS     `hcl:"cors,block"`
 	DisableAccessControl []string  `hcl:"disable_access_control,optional"`
 	Endpoints            Endpoints `hcl:"endpoint,block"`
 	ErrorFile            string    `hcl:"error_file,optional"`
-	Name                 string
+	Name                 string    `hcl:"name,label"`
 	Remain               hcl.Body  `hcl:",remain"`
 	Scope                cty.Value `hcl:"beta_scope,optional"`
 
@@ -32,6 +33,17 @@ func (a API) HCLBody() hcl.Body {
 	return a.Remain
 }
 
+// Inline implements the <Inline> interface.
+func (a API) Inline() interface{} {
+	type Inline struct {
+		AddResponseHeaders map[string]string `hcl:"add_response_headers,optional"`
+		DelResponseHeaders []string          `hcl:"remove_response_headers,optional"`
+		SetResponseHeaders map[string]string `hcl:"set_response_headers,optional"`
+	}
+
+	return &Inline{}
+}
+
 // Schema implements the <Inline> interface.
 func (a API) Schema(inline bool) *hcl.BodySchema {
 	if !inline {
@@ -39,13 +51,7 @@ func (a API) Schema(inline bool) *hcl.BodySchema {
 		return schema
 	}
 
-	type Inline struct {
-		AddResponseHeaders map[string]string `hcl:"add_response_headers,optional"`
-		DelResponseHeaders []string          `hcl:"remove_response_headers,optional"`
-		SetResponseHeaders map[string]string `hcl:"set_response_headers,optional"`
-	}
-
-	schema, _ := gohcl.ImpliedBodySchema(&Inline{})
+	schema, _ := gohcl.ImpliedBodySchema(a.Inline())
 
 	return schema
 }