+++ config/endpoint.go
@@ -12,6 +12,7 @@ var _ Inline = &Endpoint{}
 
 // Endpoint represents the <Endpoint> object.
 type Endpoint struct {
+	ErrorHandlerSetter
 	AccessControl        []string  `hcl:"access_control,optional"`
 	DisableAccessControl []string  `hcl:"disable_access_control,optional"`
 	ErrorFile            string    `hcl:"error_file,optional"`
@@ -34,13 +35,8 @@ func (e Endpoint) HCLBody() hcl.Body {
 	return e.Remain
 }
 
-// Schema implements the <Inline> interface.
-func (e Endpoint) Schema(inline bool) *hcl.BodySchema {
-	if !inline {
-		schema, _ := gohcl.ImpliedBodySchema(e)
-		return schema
-	}
-
+// Inline implements the <Inline> interface.
+func (e Endpoint) Inline() interface{} {
 	type Inline struct {
 		meta.Attributes
 		Proxies        Proxies  `hcl:"proxy,block"`
@@ -48,7 +44,17 @@ func (e Endpoint) Schema(inline bool) *hcl.BodySchema {
 		ResponseStatus *uint8   `hcl:"set_response_status,optional"`
 	}
 
-	schema, _ := gohcl.ImpliedBodySchema(&Inline{})
+	return &Inline{}
+}
 
-	return schema
+// Schema implements the <Inline> interface.
+func (e Endpoint) Schema(inline bool) *hcl.BodySchema {
+	if !inline {
+		schema, _ := gohcl.ImpliedBodySchema(e)
+		return schema
+	}
+
+	schema, _ := gohcl.ImpliedBodySchema(e.Inline())
+
+	return meta.SchemaWithAttributes(schema)
 }