+++ server/http_integration_test.go
@@ -24,6 +24,7 @@ import (
 	"time"
 
 	"github.com/dgrijalva/jwt-go/v4"
+	"github.com/google/go-cmp/cmp"
 	"github.com/sirupsen/logrus"
 	logrustest "github.com/sirupsen/logrus/hooks/test"
 
@@ -81,7 +82,7 @@ func teardown() {
 	testBackend.Close()
 }
 func newCouper(file string, helper *test.Helper) (func(), *logrustest.Hook) {
-	couperConfig, err := configload.LoadFile(filepath.Join(testWorkingDir, file))
+	couperConfig, err := configload.LoadFile(filepath.Join(testWorkingDir, file), false)
 	helper.Must(err)
 
 	return newCouperWithConfig(couperConfig, helper)
@@ -109,7 +110,7 @@ func newCouperWithTemplate(file string, helper *test.Helper, vars map[string]int
 }
 
 func newCouperWithBytes(file []byte, helper *test.Helper) (func(), *logrustest.Hook) {
-	couperConfig, err := configload.LoadBytes(file, "couper-bytes.hcl")
+	couperConfig, err := configload.LoadBytes(file, "couper-bytes.hcl", false)
 	helper.Must(err)
 
 	return newCouperWithConfig(couperConfig, helper)
@@ -2544,7 +2545,7 @@ func TestHTTPServer_backend_requests_variables(t *testing.T) {
 			http.Header{},
 			strings.NewReader(`abcd1234`),
 			expectation{
-				Method:   "POST",
+				Method:   http.MethodPost,
 				Protocol: resourceOrigin.Scheme,
 				Host:     resourceOrigin.Hostname(),
 				Port:     port,
@@ -2563,7 +2564,7 @@ func TestHTTPServer_backend_requests_variables(t *testing.T) {
 			http.Header{"Content-Type": []string{"application/json"}},
 			strings.NewReader(`{"s":"abcd1234"}`),
 			expectation{
-				Method:   "POST",
+				Method:   http.MethodPost,
 				Protocol: resourceOrigin.Scheme,
 				Host:     resourceOrigin.Hostname(),
 				Port:     port,
@@ -2582,7 +2583,7 @@ func TestHTTPServer_backend_requests_variables(t *testing.T) {
 			http.Header{"Content-Type": []string{"application/x-www-form-urlencoded"}},
 			strings.NewReader(`s=abcd1234`),
 			expectation{
-				Method:   "POST",
+				Method:   http.MethodPost,
 				Protocol: resourceOrigin.Scheme,
 				Host:     resourceOrigin.Hostname(),
 				Port:     port,
@@ -3135,7 +3136,7 @@ func TestJWTAccessControl(t *testing.T) {
 
 func TestJWTAccessControlSourceConfig(t *testing.T) {
 	helper := test.New(t)
-	couperConfig, err := configload.LoadFile("testdata/integration/config/05_couper.hcl")
+	couperConfig, err := configload.LoadFile("testdata/integration/config/05_couper.hcl", false)
 	helper.Must(err)
 
 	log, _ := logrustest.NewNullLogger()
@@ -3160,13 +3161,14 @@ func TestJWTAccessControl_round(t *testing.T) {
 	defer shutdown()
 
 	type testCase struct {
-		name string
-		path string
+		name      string
+		path      string
+		expGroups []interface{}
 	}
 
 	for _, tc := range []testCase{
-		{"separate jwt_signing_profile/jwt", "/separate"},
-		{"self-signed jwt", "/self-signed"},
+		{"separate jwt_signing_profile/jwt", "/separate", []interface{}{"g1", "g2"}},
+		{"self-signed jwt", "/self-signed", []interface{}{}},
 	} {
 		t.Run(tc.path, func(subT *testing.T) {
 			helper := test.New(subT)
@@ -3217,6 +3219,17 @@ func TestJWTAccessControl_round(t *testing.T) {
 			if pidclaim != pid {
 				subT.Fatalf("%q: unexpected pid claim: %q", tc.name, pidclaim)
 			}
+			groupsclaim, ok := claims["groups"]
+			if !ok {
+				subT.Fatalf("%q: missing groups claim: %#v", tc.name, claims)
+			}
+			groupsclaimArray, ok := groupsclaim.([]interface{})
+			if !ok {
+				subT.Fatalf("%q: groups must be array: %#v", tc.name, groupsclaim)
+			}
+			if !cmp.Equal(tc.expGroups, groupsclaimArray) {
+				subT.Errorf(cmp.Diff(tc.expGroups, groupsclaimArray))
+			}
 		})
 	}
 }
@@ -3258,15 +3271,15 @@ func Test_Scope(t *testing.T) {
 	for _, tc := range []testCase{
 		{"by scope: unauthorized", http.MethodGet, "/scope/foo", false, http.StatusUnauthorized, "access control error: scoped_jwt: token required", "jwt_token_missing"},
 		{"by scope: sufficient scope", http.MethodGet, "/scope/foo", true, http.StatusNoContent, "", ""},
-		{"by scope: additional scope required: insufficient scope", http.MethodPost, "/scope/foo", true, http.StatusForbidden, `access control error: scope: required scope "foo" not granted`, "beta_insufficient_scope"},
-		{"by scope: operation not permitted", http.MethodDelete, "/scope/foo", true, http.StatusForbidden, "access control error: scope: operation DELETE not permitted", "beta_operation_denied"},
-		{"by scope: additional scope required by *: insufficient scope", http.MethodGet, "/scope/bar", true, http.StatusForbidden, `access control error: scope: required scope "more" not granted`, "beta_insufficient_scope"},
+		{"by scope: additional scope required: insufficient scope", http.MethodPost, "/scope/foo", true, http.StatusForbidden, `access control error: required scope "foo" not granted`, "beta_insufficient_scope"},
+		{"by scope: operation not permitted", http.MethodDelete, "/scope/foo", true, http.StatusForbidden, "access control error: operation DELETE not permitted", "beta_operation_denied"},
+		{"by scope: additional scope required by *: insufficient scope", http.MethodGet, "/scope/bar", true, http.StatusForbidden, `access control error: required scope "more" not granted`, "beta_insufficient_scope"},
 		{"by scope: no additional scope required: sufficient scope", http.MethodDelete, "/scope/bar", true, http.StatusNoContent, "", ""},
 		{"by role: unauthorized", http.MethodGet, "/role/foo", false, http.StatusUnauthorized, "access control error: roled_jwt: token required", "jwt_token_missing"},
 		{"by role: sufficient scope", http.MethodGet, "/role/foo", true, http.StatusNoContent, "", ""},
-		{"by role: additional scope required: insufficient scope", http.MethodPost, "/role/foo", true, http.StatusForbidden, `access control error: scope: required scope "foo" not granted`, "beta_insufficient_scope"},
-		{"by role: operation not permitted", http.MethodDelete, "/role/foo", true, http.StatusForbidden, "access control error: scope: operation DELETE not permitted", "beta_operation_denied"},
-		{"by role: additional scope required by *: insufficient scope", http.MethodGet, "/role/bar", true, http.StatusForbidden, `access control error: scope: required scope "more" not granted`, "beta_insufficient_scope"},
+		{"by role: additional scope required: insufficient scope", http.MethodPost, "/role/foo", true, http.StatusForbidden, `access control error: required scope "foo" not granted`, "beta_insufficient_scope"},
+		{"by role: operation not permitted", http.MethodDelete, "/role/foo", true, http.StatusForbidden, "access control error: operation DELETE not permitted", "beta_operation_denied"},
+		{"by role: additional scope required by *: insufficient scope", http.MethodGet, "/role/bar", true, http.StatusForbidden, `access control error: required scope "more" not granted`, "beta_insufficient_scope"},
 		{"by role: no additional scope required: sufficient scope", http.MethodDelete, "/role/bar", true, http.StatusNoContent, "", ""},
 	} {
 		t.Run(fmt.Sprintf("%s_%s_%s", tc.name, tc.operation, tc.path), func(subT *testing.T) {